requirements.txt
Django==1.9.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
gunicorn==19.4.5
markdown2==2.3.0
Pillow==3.1.0
psycopg2==2.6.1
sorl-thumbnail==12.3
wheel==0.26.0
